refactor(admin): split users.ts into focused modules
902行の admin/users.ts を機能単位のモジュールに分割し、
1,000行の健全性境界を下回る構成に整理した。

- types.ts: 共有型定義
- list.ts: getAllUsers
- detail.ts: getUserDetail
- koudens.ts: getAllKoudens
- auth-info.ts / helpers.ts: 内部ヘルパー
- index.ts: 既存 import パス維持のための re-export

Closes #133
